pyproject.toml
[project]
name = "mega-sistema-ia"
version = "1.0.0"
description = "AI voice agent template for any business — powered by Retell AI, Modal, and Twilio"
requires-python = ">=3.11"
dependencies = [
    "modal>=0.73.0",
    "retell-sdk>=5.0.0",
    "twilio>=9.0.0",
    "anthropic>=0.42.0",
    "notion-client>=2.2.0,<3.0.0",
    "requests>=2.32.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "pytest>=8.0.0",
]
